+++ CHANGELOG.md
@@ -8,6 +8,7 @@
     - Upgraded Azure provider to 2.99.0
   - Add `variables_file` option to the `mach_composer` configuration block to define a variable file
   - Fix auto add cloud integration (aws or azure) when `integration` list is left empty
+  - Add ability to define a custom provider version including the version operator
 - Deprecate `commercetools.frontend` block, will be removed in a later release.
 
 ## 2.1.1 (2022-04-22)